added safety net for objects that fall off the map

+++ Digitakomo/Assets/Script/Enemies/Shared/EnemyBaseClass.cs
@@ -144,7 +144,7 @@ public abstract class EnemyBaseClass : MonoBehaviour
         }
     }
     // Returns Hp
-    public float GetCurrentHP()
+    public int GetCurrentHP()
     {
         return hp;
     }
